move helpers from ShopkeeperApp.tsx to their own files
